This adds a new bpf map cilium_lb4_backends_v2
The V2 map key is u32 typed instead of u16.

Upon agent (re)start:
- we restore backends from v1 map and copy all
entries from v1 map to v2 map if v1 map exists
(this is the upgrade scenario).
- we then remove v1 map and operate on v2 map.

+++ pkg/service/service_test.go
@@ -339,7 +339,7 @@ func (m *ManagerTestSuite) TestRestoreServices(c *C) {
 	c.Assert(len(matches), Equals, 1) // only the id2 svc has session affinity
 	c.Assert(len(matches[uint16(id2)]), Equals, 2)
 	for _, b := range lbmap.ServiceByID[uint16(id2)].Backends {
-		c.Assert(m.lbmap.AffinityMatch[uint16(id1)][uint16(b.ID)], Equals, struct{}{})
+		c.Assert(m.lbmap.AffinityMatch[uint16(id1)][b.ID], Equals, struct{}{})
 	}
 
 }
